mangsir23

+++ backend/utils/removeimg.js
@@ -1,4 +1,11 @@
 import fs from "fs";
-export const removeImg = (path) => {
-  fs.unlinkSync(path);
+
+export const removeimg = (path) => {
+  try {
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+    }
+  } catch (error) {
+    console.error("Error removing image:", error);
+  }
 };
